requirements.txt
Flask
requests
json
python-google-places==1.1.0

